cleanup(coral-tokens): renaming shades transform
